backend/requirements.txt
Django==3.2.20
djangorestframework==3.14.0
djongo==1.3.6
pymongo==3.12.3
sqlparse==0.2.4
django-cors-headers==4.3.1
python-dotenv==1.0.0
PyJWT==2.8.0
